ref: move logic from controller to service VideoStorageService
